Refactor non_max_suppression function into nms.py
